facts & congrats pages separated outed
