+++ packages/app/src/pages/CountWithMe/CountWithMe.tsx
@@ -1,6 +1,7 @@
-import { IonButton, IonText } from "@ionic/react";
-import React, { useState, useEffect } from "react";
+import { CongratsPage } from "./CountCongrats";
+import { FactsPage } from "./CountFacts";
 import { FormattedMessage } from "react-intl";
+import { IonButton, IonText } from "@ionic/react";
 
 interface Animal {
   url: string;
@@ -77,168 +78,6 @@ const FirstPage = (animalIndex: number, data: DataItem[]) => {
               />
             ),
           )}
-        </div>
-      </div>
-    </>
-  );
-};
-
-const FactsPage = (animalIndex: number, data: DataItem[]) => {
-  // Function to render the facts page for each animal
-  return (
-    <>
-      <div
-        style={{
-          backgroundColor: "#F7FAF9",
-          display: "flex",
-          justifyContent: "center",
-          alignItems: "center",
-          height: "100vh",
-        }}
-      >
-        <div
-          style={{
-            position: "relative",
-            backgroundColor: "#FFFFFF",
-            borderRadius: "20px",
-            width: "1159px",
-            height: "640px",
-            boxShadow: "0 8px 16px rgba(0, 0, 0, 0.2)",
-            display: "flex",
-            flexDirection: "row", // Use row to align content horizontally
-            justifyContent: "center", // Center content horizontally
-            alignItems: "stretch", // Align items to stretch vertically
-            padding: "20px",
-          }}
-        >
-          <div
-            style={{
-              flex: 1,
-              display: "flex",
-              flexDirection: "column",
-              justifyContent: "center",
-              alignItems: "flex-start", // Align text content to the left
-              padding: "20px",
-            }}
-          >
-            <IonText>
-              <h2>{data[animalIndex].fact[0].text}</h2>
-              <p>{data[animalIndex].fact[1].text}</p>
-            </IonText>
-          </div>
-
-          <div
-            style={{
-              flex: 1,
-              display: "flex",
-              flexDirection: "column",
-              justifyContent: "center",
-              alignItems: "flex-end", // Align map image to the right
-              padding: "20px",
-            }}
-          >
-            <img
-              src={data[animalIndex].map}
-              alt={`animal-${animalIndex}`}
-              style={{ width: "100%", height: "100%" }}
-            />
-          </div>
-        </div>
-      </div>
-    </>
-  );
-};
-
-const CongratsPage = () => {
-  // Function to render the congrats page
-  const congrats = {
-    background:
-      "https://ik.imagekit.io/jskeetedev/Untitled%20design%20(3).png?updatedAt=1706831646016",
-    star: "https://ik.imagekit.io/jskeetedev/Untitled%20design%20(2).png?updatedAt=1706831320447",
-    es: "Te ganaste una estrella",
-    en: "You've earned a star",
-  };
-
-  const [showText, setShowText] = useState(true); // State to show/hide text
-
-  useEffect(() => {
-    const timeout = setTimeout(() => {
-      setShowText(false);
-    }, 2000); // Set timeout to hide text after 2 seconds
-
-    return () => clearTimeout(timeout);
-  }, []); // This effect runs only once
-
-  return (
-    <>
-      <div style={{ padding: "50px" }}>
-        <div
-          style={{
-            height: "100vh",
-            display: "flex",
-            flexDirection: "column",
-            justifyContent: "center",
-            alignItems: "center",
-            position: "relative",
-          }}
-        >
-          <img
-            src={congrats.background}
-            style={{
-              position: "absolute",
-              width: "1200px",
-              height: "auto",
-              zIndex: 1,
-            }}
-            alt="background"
-          />
-          <IonText>
-            <h1>Actividad Completada</h1>
-            <p style={{ textAlign: "center" }}>Activity Completed</p>
-          </IonText>
-
-          <div
-            style={{
-              position: "relative",
-              zIndex: 2,
-              textAlign: "center",
-            }}
-          >
-            <img
-              src={congrats.star}
-              alt="star"
-              style={{
-                width: showText ? "700px" : "200px",
-                height: "auto",
-                position: "relative",
-                transition: "width 1s ease", // transition effect to animate star shrink
-              }}
-            />
-            {showText && (
-              <div
-                style={{
-                  position: "absolute",
-                  top: "50%",
-                  left: "50%",
-                  transform: "translate(-50%, -50%)", // Center text horizontally and vertically in the star
-                  color: "#fff",
-                }}
-              >
-                <p style={{ fontSize: "26px", margin: 0, color: "black" }}>
-                  {congrats.es}
-                </p>
-                <p style={{ fontSize: "12px", margin: 0, color: "black" }}>
-                  {congrats.en}
-                </p>
-              </div>
-            )}
-          </div>
-
-          <IonButton shape="round" style={{ width: "300px" }}>
-            ¡Sigue adelante!
-            <br />
-            Keep Going!
-          </IonButton>
         </div>
       </div>
     </>
